Docker/audio_qa_api/requirements.txt
fastapi==0.95.2
gunicorn==20.1.0
uvicorn==0.17.6
protobuf==3.20.*
tqdm==4.64.0
PyOpenGL==3.1.6
pickle5==0.0.11
pandas==1.4.3
imageio==2.22.2
ffmpeg-python==0.2.0
python-multipart==0.0.5
moviepy==1.0.3
openai==0.27.6
llama-index==0.4.23
langchain==0.0.299
chromadb==0.3.21
tiktoken==0.3.3
graphsignal==0.13.26
redis==4.5.5
azure-identity==1.13.0
azure-search-documents==11.4.0b8